+++ src/index.ts
@@ -1,10 +1,13 @@
 
 import { Task, ProvingTask, DeployTask, QueryParams, VerifyData, StatusState, DeploymentInfo, Image, Statistics, AddWasmImageTask } from "./interface/interface.js";
 
-import { ZkWasmServiceHelper, ZkWasmServiceImageHelper, ZkWasmServiceTaskHelper, ZkWasmUtil } from "./helper/helper.js";
+import { ZkWasmUtil } from "helper/util.js";
+import { ZkWasmServiceEndpoint } from "helper/endpoint.js";
+import { ZkWasmServiceImageHelper } from "helper/image.js";
+import { ZkWasmServiceTaskHelper } from "helper/task.js";
 
 export {
     Task, ProvingTask, DeployTask, QueryParams, VerifyData, StatusState, DeploymentInfo, Image, Statistics, AddWasmImageTask,
-    ZkWasmServiceHelper, ZkWasmServiceImageHelper, ZkWasmServiceTaskHelper, ZkWasmUtil
+    ZkWasmServiceEndpoint, ZkWasmServiceImageHelper, ZkWasmServiceTaskHelper, ZkWasmUtil
 };
